Moved part of list index endpoints to router
